refactor: add useMonthRangeHeader hook
